Create helper module for fetching JSON data
This avoids a small amount of repeated code, and makes for a nicer
place to put the Json type definition.
